+++ run_client.sh
@@ -11,3 +11,4 @@ gnome-terminal -- python3 SimilarityConsumer.py
 gnome-terminal -- python3 SizeFilterConsumer.py 
 gnome-terminal -- python3 FacesFilterConsumer.py 
 gnome-terminal -- python3 App.py 
+gnome-terminal -- cd ../text_server && mix run ".\lib\receive.exs" --no-halt
