+++ text_server/README.md
@@ -1,21 +1,12 @@
 # TextServer
 
-**TODO: Add description**
+Text server is a component that detects text on images.
 
 ## Installation
 
-If [available in Hex](https://hex.pm/docs/publish), the package can be installed
-by adding `text_server` to your list of dependencies in `mix.exs`:
-
 ```elixir
-def deps do
-  [
-    {:text_server, "~> 0.1.0"}
-  ]
-end
+> mix deps.get
+> mix compile
+> mix run  --no-halt
+> mix run .\lib\receive.exs --no-halt
 ```
-
-Documentation can be generated with [ExDoc](https://github.com/elixir-lang/ex_doc)
-and published on [HexDocs](https://hexdocs.pm). Once published, the docs can
-be found at <https://hexdocs.pm/text_server>.
-
