+++ text_server/lib/text_server.ex
@@ -1,17 +1,4 @@
 defmodule TextServer do
-  @moduledoc """
-  Documentation for `TextServer`.
-  """
-
-  @doc """
-  Hello world.
-
-  ## Examples
-
-      iex> TextServer.hello()
-      :world
-
-  """
   def hello do
     :world
   end
